+++ R/if-null.R
@@ -6,13 +6,14 @@
 #' A mostly copy of `rlang`'s `%||%` except does not use `rlang::is_null()`,
 #'   which, currently, calls the same primitive [base::is.null] function.
 #'
+#'   Note: `%||%` is copied from `{base}` if available (**R** versions >= 4.4)
+#'
 #' @param x,y If `x` is `NULL` returns `y`; otherwise `x`
 #'
 #' @name if_null
 #' @return `x` if it is not `NULL` or has length, depending on check
-#'
 #' @examples
-#' # replace NULL
+#' # replace NULL (for R < 4.4)
 #' NULL %||% 1L
 #' 2L   %||% 1L
 #'
@@ -24,11 +25,14 @@
 #'
 #' # replace no length
 #' logical() %len% TRUE
-#" FALSE     %len% TRUE
+#' FALSE     %len% TRUE
+NULL
+
+#' @rdname if_null
 #' @export
-`%||%` <- function(x, y) {
+`%||%` <- get0("%||%", baseenv(), "function", ifnotfound = function(x, y) {
   if (is.null(x)) y else x
-}
+})
 
 #' @rdname if_null
 #' @export